добавил сущности и маппер для работы с бд

+++ app/src/main/java/com/example/pizza_shift_2025/data/model/DoughModel.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class DoughModel(
     val name: DoughTypeModel,
-    val price: Double
+    val price: Int
 )
